hotel_api/requirements.txt
appdirs==1.4.3
Django==1.10.6
django-cors-headers==2.1.0
django-filter==1.0.2
django-location-field==2.0.3
djangorestframework==3.6.2
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.2.0
six==1.10.0
